Added files via upload.
Suroy is a device that aids jeepney riders to track jeepneys through the GPS modules of the mobile phones of drivers, and allows the driver to track the passenger count of the rider as well. The Suroy Cube is an Arduino based project that accepts button inputs intended for jeepney passengers and sends a Bluetooth signal to the driver's phone app to track the passenger count.